moved helpers to utils and removed memes helper.
i don't love this, but it was a pretty useless helper file anyway. utils and helpers are both commonly used in python, so either works. but i figure we'll go with utils instead.
