+++ RyuZU/Core/utils.py
